Improve Http4sDsl
