--- dsl/src/main/scala/org/http4s/dsl/Auth.scala
@@ -1,7 +1,0 @@
-package org.http4s
-package dsl
-
-object as {
-  def unapply[F[_], A](ar: AuthedRequest[F, A]): Option[(Request[F], A)] =
-    Some(ar.req -> ar.authInfo)
-}
